refactor(jazz-tools): Phase 8 (routes) — split server/routes.rs into http/websocket/utils submodules
server/routes.rs was 3.6K LOC mixing HTTP endpoint handlers, WebSocket
lifecycle, request/response DTOs, and parser helpers. Split into a
directory with three submodules + a thin re-exporting root:

  server/routes/
    mod.rs       — module docs, create_router (re-exports unchanged so
                   server::routes::create_router still resolves), tests
    http.rs      — HTTP endpoint handlers + request/response DTOs +
                   forward_catalogue_request/authority_endpoint_url
    websocket.rs — ws_handler, handshake authentication
                   (authenticate_ws_handshake, request_uses_cookie_auth,
                   validate_ws_cookie_origin, is_loopback_*),
                   send_ws_error, handle_ws_connection, ws_cleanup
    utils.rs     — shared parser/validator helpers (parse_schema_hash_param,
                   parse_object_id_param, parse_app_id_param,
                   permissions_head_view, permissions_map_view,
                   unix_timestamp_millis, connection_schema_diagnostics_from_handshake)

Cross-file items that need to be visible to siblings are exposed as
pub(super) — handlers, types, and the WS handshake outcome enum
WsClientSetup. The PermissionsHeadView struct (constructed in utils.rs,
serialized in http.rs) gains pub(super) on its fields. SchemaConnectivityParams
gains pub(super) on its fields for the same reason.

The pre-existing tests block stays in mod.rs and uses
`use super::http::*; use super::utils::*; use super::websocket::*;`
plus a small targeted set of stdlib/axum imports to keep the test
helpers' references resolving without changing any test logic.

Independent of every other phase. Spec:
specs/todo/c_later/crate_structure_cleanup.md (Phase 8 — routes half;
the row_histories half landed earlier in PR #712).
