+++ crates/jazz-tools/src/server/routes/mod.rs
@@ -1,32 +1,35 @@
-//! HTTP routes for the Jazz server.
+//! HTTP and WebSocket routes for the Jazz server.
+//!
+//! Split into three submodules so each piece is independently navigable:
+//! - [`http`] — HTTP endpoint handlers and their request/response types
+//! - [`websocket`] — WebSocket lifecycle (handshake auth, connection, cleanup)
+//! - [`utils`] — parser/validator helpers used by both
+//!
+//! The router builder [`create_router`] re-exports unchanged from this module
+//! so existing callers (`server::routes::create_router`) continue to resolve.
+
+mod http;
+mod utils;
+mod websocket;
 
 use std::sync::Arc;
-use std::time::{SystemTime, UNIX_EPOCH};
 
 use axum::{
     Router,
-    extract::ws::{Message, WebSocket, WebSocketUpgrade},
-    extract::{Path, Query, State},
-    http::{HeaderMap, StatusCode, header::CONTENT_TYPE},
-    response::{IntoResponse, Json, Response},
     routing::{get, post},
 };
-use serde::{Deserialize, Serialize};
 use tower_http::cors::CorsLayer;
 use tower_http::trace::TraceLayer;
-use uuid::Uuid;
 
-use crate::jazz_transport::ErrorResponse;
-use crate::middleware::auth::{extract_session, validate_admin_secret, validate_backend_secret};
-use crate::object::ObjectId;
-use crate::query_manager::types::{
-    ColumnType, Schema, SchemaHash, TableName, TablePolicies, Value,
+use crate::server::ServerState;
+
+use http::{
+    admin_subscription_introspection_handler, health_handler, permissions_handler,
+    permissions_head_handler, publish_migration_handler, publish_permissions_handler,
+    publish_schema_handler, schema_connectivity_handler, schema_handler, schema_hashes_handler,
 };
-use crate::schema_manager::{AppId, Lens, LensOp, LensTransform};
-use crate::server::{CatalogueAuthorityMode, ConnectionState, ServerState};
-use crate::sync_manager::ClientId;
+use websocket::ws_handler;
 
-/// Create the router with all routes.
 pub fn create_router(state: Arc<ServerState>) -> Router {
     // TODO: Accept app-name aliases in app-scoped route matching
     // Nesting all non-health routes under a fixed "/apps/{state.app_id}" path makes the server only match the canonical UUID string, but JavaScript callers frequently propagate human-readable app IDs (e.g. "test-app") that are valid elsewhere via AppId::from_name(...). In that non-UUID case, the client now builds /apps/test-app/... URLs while the server only serves /apps/<derived-uuid>/..., so websocket and admin/schema requests return 404 for otherwise valid app IDs.
@@ -59,1633 +62,23 @@ pub fn create_router(state: Arc<ServerState>) -> Router {
         .with_state(state)
 }
 
-#[derive(Debug, Serialize)]
-struct SchemaHashesResponse {
-    hashes: Vec<String>,
-}
-
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct StoredSchemaResponse {
-    schema: Schema,
-    published_at: Option<u64>,
-}
-
-#[derive(Debug, Deserialize)]
-struct AdminSubscriptionIntrospectionParams {
-    #[serde(rename = "appId")]
-    app_id: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct SchemaConnectivityParams {
-    from_hash: String,
-    to_hash: String,
-}
-
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct AdminSubscriptionIntrospectionResponse {
-    app_id: String,
-    generated_at: u64,
-    queries: Vec<crate::query_manager::manager::ServerSubscriptionTelemetryGroup>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct PublishMigrationRequest {
-    from_hash: String,
-    to_hash: String,
-    forward: Vec<PublishTableLens>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct PublishTableLens {
-    table: String,
-    #[serde(default)]
-    added: bool,
-    #[serde(default)]
-    removed: bool,
-    renamed_from: Option<String>,
-    operations: Vec<PublishLensOp>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(tag = "type", rename_all = "camelCase")]
-enum PublishLensOp {
-    Introduce {
-        column: String,
-        column_type: ColumnType,
-        value: Value,
-    },
-    Drop {
-        column: String,
-        column_type: ColumnType,
-        value: Value,
-    },
-    Rename {
-        column: String,
-        value: String,
-    },
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-struct PublishSchemaRequest {
-    schema: Schema,
-    permissions: Option<std::collections::HashMap<TableName, TablePolicies>>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct PublishPermissionsRequest {
-    schema_hash: String,
-    permissions: std::collections::HashMap<String, TablePolicies>,
-    expected_parent_bundle_object_id: Option<String>,
-}
-
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct PublishSchemaResponse {
-    object_id: String,
-    hash: String,
-}
-
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct PermissionsHeadView {
-    schema_hash: String,
-    version: u64,
-    parent_bundle_object_id: Option<String>,
-    bundle_object_id: String,
-}
-
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct PermissionsHeadResponse {
-    head: Option<PermissionsHeadView>,
-}
-
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct StoredPermissionsResponse {
-    head: Option<PermissionsHeadView>,
-    permissions: Option<std::collections::HashMap<String, TablePolicies>>,
-}
-
-#[derive(Debug, Serialize)]
-struct SchemaConnectivityResponse {
-    connected: bool,
-}
-
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct PublishMigrationResponse {
-    object_id: String,
-    from_hash: String,
-    to_hash: String,
-}
-
-async fn forward_catalogue_request(
-    state: &Arc<ServerState>,
-    method: reqwest::Method,
-    path: &str,
-    body: Option<Vec<u8>>,
-) -> Result<Response, (StatusCode, Json<ErrorResponse>)> {
-    let (base_url, authority_admin_secret) = match &state.catalogue_authority {
-        CatalogueAuthorityMode::Local => {
-            return Err((
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::internal(
-                    "catalogue forwarding requested without a configured authority".to_string(),
-                )),
-            ));
-        }
-        CatalogueAuthorityMode::Forward {
-            base_url,
-            admin_secret,
-        } => (base_url.as_str(), admin_secret.as_str()),
-    };
-
-    let app_scoped_path = format!("/apps/{}/{}", state.app_id, path.trim_start_matches('/'));
-    let authority_url = authority_endpoint_url(base_url, &app_scoped_path).map_err(|message| {
-        (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ErrorResponse::internal(message)),
-        )
-    })?;
-
-    let mut request = state
-        .http_client
-        .request(method, authority_url)
-        .header("X-Jazz-Admin-Secret", authority_admin_secret);
-    if let Some(body) = body {
-        request = request.header(CONTENT_TYPE, "application/json").body(body);
-    }
-
-    let response = request.send().await.map_err(|err| {
-        (
-            StatusCode::BAD_GATEWAY,
-            Json(ErrorResponse::internal(format!(
-                "failed to reach catalogue authority: {err}"
-            ))),
-        )
-    })?;
-
-    let status =
-        StatusCode::from_u16(response.status().as_u16()).unwrap_or(StatusCode::BAD_GATEWAY);
-    let content_type = response.headers().get(CONTENT_TYPE).cloned();
-    let bytes = response.bytes().await.map_err(|err| {
-        (
-            StatusCode::BAD_GATEWAY,
-            Json(ErrorResponse::internal(format!(
-                "failed to read authority response: {err}"
-            ))),
-        )
-    })?;
-
-    let mut response_builder = Response::builder().status(status);
-    if let Some(content_type) = content_type {
-        response_builder = response_builder.header(CONTENT_TYPE, content_type);
-    }
-
-    response_builder
-        .body(axum::body::Body::from(bytes))
-        .map_err(|err| {
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::internal(format!(
-                    "failed to build forwarded response: {err}"
-                ))),
-            )
-        })
-}
-
-fn authority_endpoint_url(base_url: &str, path: &str) -> Result<String, String> {
-    let parsed = reqwest::Url::parse(base_url)
-        .map_err(|err| format!("invalid catalogue authority URL '{base_url}': {err}"))?;
-    let mut origin = parsed.clone();
-    origin.set_query(None);
-    origin.set_fragment(None);
-
-    let (path_only, query) = match path.split_once('?') {
-        Some((path_only, query)) => (path_only, Some(query)),
-        None => (path, None),
-    };
-
-    let mut full_path = parsed.path().trim_end_matches('/').to_string();
-    if full_path.is_empty() {
-        full_path.push('/');
-    }
-    if !full_path.ends_with('/') {
-        full_path.push('/');
-    }
-    full_path.push_str(path_only.trim_start_matches('/'));
-
-    origin.set_path(&full_path);
-    origin.set_query(query);
-    Ok(origin.to_string())
-}
-
-/// Return the catalogue schema for the given hash plus its publish timestamp.
-///
-/// Requires a valid admin secret; returns 404 if no schema exists for the hash.
-async fn schema_handler(
-    State(state): State<Arc<ServerState>>,
-    Path(hash_text): Path<String>,
-    headers: HeaderMap,
-) -> impl IntoResponse {
-    let admin_secret = headers
-        .get("X-Jazz-Admin-Secret")
-        .and_then(|v| v.to_str().ok());
-
-    match validate_admin_secret(admin_secret, &state.auth_config) {
-        Ok(()) => {}
-        Err((status, msg)) => {
-            return (status, Json(ErrorResponse::unauthorized(msg))).into_response();
-        }
-    }
-
-    if matches!(
-        &state.catalogue_authority,
-        CatalogueAuthorityMode::Forward { .. }
-    ) {
-        return match forward_catalogue_request(
-            &state,
-            reqwest::Method::GET,
-            &format!("/schema/{hash_text}"),
-            None,
-        )
-        .await
-        {
-            Ok(response) => response,
-            Err(error) => error.into_response(),
-        };
-    }
-
-    let schema_hash = match parse_schema_hash_param(&hash_text) {
-        Ok(hash) => hash,
-        Err(message) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ErrorResponse::bad_request(message)),
-            )
-                .into_response();
-        }
-    };
-
-    match state.runtime.known_schema(&schema_hash) {
-        Ok(Some(schema)) => {
-            let published_at = match state.runtime.schema_published_at(&schema_hash) {
-                Ok(timestamp) => timestamp,
-                Err(err) => {
-                    return (
-                        StatusCode::INTERNAL_SERVER_ERROR,
-                        Json(ErrorResponse::internal(format!(
-                            "failed to read schema publish timestamp: {err}"
-                        ))),
-                    )
-                        .into_response();
-                }
-            };
-            tracing::info!(
-                requested_hash = %schema_hash.short(),
-                "schema request: returning requested hash"
-            );
-            let body = StoredSchemaResponse {
-                schema: schema.clone(),
-                published_at,
-            };
-            Json(body).into_response()
-        }
-        Ok(None) => (
-            StatusCode::NOT_FOUND,
-            Json(ErrorResponse::not_found(format!(
-                "schema catalogue not found for hash {}",
-                schema_hash
-            ))),
-        )
-            .into_response(),
-        Err(err) => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ErrorResponse::internal(format!(
-                "failed to read schema catalogue: {err}"
-            ))),
-        )
-            .into_response(),
-    }
-}
-
-/// Return all known schema hashes from catalogue state.
-///
-/// Requires a valid admin secret.
-async fn schema_hashes_handler(
-    State(state): State<Arc<ServerState>>,
-    headers: HeaderMap,
-) -> impl IntoResponse {
-    let admin_secret = headers
-        .get("X-Jazz-Admin-Secret")
-        .and_then(|v| v.to_str().ok());
-
-    match validate_admin_secret(admin_secret, &state.auth_config) {
-        Ok(()) => {}
-        Err((status, msg)) => {
-            return (status, Json(ErrorResponse::unauthorized(msg))).into_response();
-        }
-    }
-
-    if matches!(
-        &state.catalogue_authority,
-        CatalogueAuthorityMode::Forward { .. }
-    ) {
-        return match forward_catalogue_request(&state, reqwest::Method::GET, "/schemas", None).await
-        {
-            Ok(response) => response,
-            Err(error) => error.into_response(),
-        };
-    }
-
-    match state.runtime.known_schema_hashes() {
-        Ok(hashes) => {
-            let body = SchemaHashesResponse {
-                hashes: hashes.iter().map(ToString::to_string).collect(),
-            };
-            Json(body).into_response()
-        }
-        Err(err) => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ErrorResponse::internal(format!(
-                "failed to read schema hashes: {err}"
-            ))),
-        )
-            .into_response(),
-    }
-}
-
-/// Return whether two known schema hashes are connected by non-draft uploaded migrations.
-///
-/// Requires a valid admin secret.
-async fn schema_connectivity_handler(
-    State(state): State<Arc<ServerState>>,
-    headers: HeaderMap,
-    Query(params): Query<SchemaConnectivityParams>,
-) -> impl IntoResponse {
-    let admin_secret = headers
-        .get("X-Jazz-Admin-Secret")
-        .and_then(|v| v.to_str().ok());
-
-    match validate_admin_secret(admin_secret, &state.auth_config) {
-        Ok(()) => {}
-        Err((status, msg)) => {
-            return (status, Json(ErrorResponse::unauthorized(msg))).into_response();
-        }
-    }
-
-    if matches!(
-        &state.catalogue_authority,
-        CatalogueAuthorityMode::Forward { .. }
-    ) {
-        let forwarded_path = format!(
-            "/admin/schema-connectivity?fromHash={}&toHash={}",
-            params.from_hash, params.to_hash
-        );
-        return match forward_catalogue_request(&state, reqwest::Method::GET, &forwarded_path, None)
-            .await
-        {
-            Ok(response) => response,
-            Err(error) => error.into_response(),
-        };
-    }
-
-    let from_hash = match parse_schema_hash_param(&params.from_hash) {
-        Ok(hash) => hash,
-        Err(message) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ErrorResponse::bad_request(message)),
-            )
-                .into_response();
-        }
-    };
-    let to_hash = match parse_schema_hash_param(&params.to_hash) {
-        Ok(hash) => hash,
-        Err(message) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ErrorResponse::bad_request(message)),
-            )
-                .into_response();
-        }
-    };
-
-    match state.runtime.with_schema_manager(|schema_manager| {
-        schema_manager.are_schema_hashes_connected(from_hash, to_hash)
-    }) {
-        Ok(connected) => (
-            StatusCode::OK,
-            Json(SchemaConnectivityResponse { connected }),
-        )
-            .into_response(),
-        Err(err) => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ErrorResponse::internal(format!(
-                "failed to compute schema connectivity: {err}"
-            ))),
-        )
-            .into_response(),
-    }
-}
-
-/// Publish a schema object into the catalogue.
-///
-/// Requires a valid admin secret.
-async fn publish_schema_handler(
-    State(state): State<Arc<ServerState>>,
-    headers: HeaderMap,
-    Json(request): Json<PublishSchemaRequest>,
-) -> impl IntoResponse {
-    let admin_secret = headers
-        .get("X-Jazz-Admin-Secret")
-        .and_then(|v| v.to_str().ok());
-
-    match validate_admin_secret(admin_secret, &state.auth_config) {
-        Ok(()) => {}
-        Err((status, msg)) => {
-            return (status, Json(ErrorResponse::unauthorized(msg))).into_response();
-        }
-    }
-
-    if matches!(
-        &state.catalogue_authority,
-        CatalogueAuthorityMode::Forward { .. }
-    ) {
-        let body = match serde_json::to_vec(&request) {
-            Ok(body) => body,
-            Err(err) => {
-                return (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    Json(ErrorResponse::internal(format!(
-                        "failed to serialize schema publish request: {err}"
-                    ))),
-                )
-                    .into_response();
-            }
-        };
-        return match forward_catalogue_request(
-            &state,
-            reqwest::Method::POST,
-            "/admin/schemas",
-            Some(body),
-        )
-        .await
-        {
-            Ok(response) => response,
-            Err(error) => error.into_response(),
-        };
-    }
-
-    if request.permissions.is_some() {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(ErrorResponse::bad_request(
-                "schema publishing no longer accepts permissions; publish permissions via POST /admin/permissions".to_string(),
-            )),
-        )
-            .into_response();
-    }
-
-    let schema_hash = SchemaHash::compute(&request.schema);
-    let object_id = match state.runtime.publish_schema(request.schema) {
-        Ok(object_id) => object_id,
-        Err(err) => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::internal(format!(
-                    "failed to publish schema catalogue: {err}"
-                ))),
-            )
-                .into_response();
-        }
-    };
-
-    (
-        StatusCode::CREATED,
-        Json(PublishSchemaResponse {
-            object_id: object_id.to_string(),
-            hash: schema_hash.to_string(),
-        }),
-    )
-        .into_response()
-}
-
-async fn permissions_head_handler(
-    State(state): State<Arc<ServerState>>,
-    headers: HeaderMap,
-) -> impl IntoResponse {
-    let admin_secret = headers
-        .get("X-Jazz-Admin-Secret")
-        .and_then(|v| v.to_str().ok());
-
-    match validate_admin_secret(admin_secret, &state.auth_config) {
-        Ok(()) => {}
-        Err((status, msg)) => {
-            return (status, Json(ErrorResponse::unauthorized(msg))).into_response();
-        }
-    }
-
-    if matches!(
-        &state.catalogue_authority,
-        CatalogueAuthorityMode::Forward { .. }
-    ) {
-        return match forward_catalogue_request(
-            &state,
-            reqwest::Method::GET,
-            "/admin/permissions/head",
-            None,
-        )
-        .await
-        {
-            Ok(response) => response,
-            Err(error) => error.into_response(),
-        };
-    }
-
-    match state.runtime.with_schema_manager(|schema_manager| {
-        schema_manager
-            .current_permissions_head()
-            .map(permissions_head_view)
-    }) {
-        Ok(head) => (StatusCode::OK, Json(PermissionsHeadResponse { head })).into_response(),
-        Err(err) => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ErrorResponse::internal(format!(
-                "failed to read permissions head: {err}"
-            ))),
-        )
-            .into_response(),
-    }
-}
-
-async fn permissions_handler(
-    State(state): State<Arc<ServerState>>,
-    headers: HeaderMap,
-) -> impl IntoResponse {
-    let admin_secret = headers
-        .get("X-Jazz-Admin-Secret")
-        .and_then(|v| v.to_str().ok());
-
-    match validate_admin_secret(admin_secret, &state.auth_config) {
-        Ok(()) => {}
-        Err((status, msg)) => {
-            return (status, Json(ErrorResponse::unauthorized(msg))).into_response();
-        }
-    }
-
-    if matches!(
-        &state.catalogue_authority,
-        CatalogueAuthorityMode::Forward { .. }
-    ) {
-        return match forward_catalogue_request(
-            &state,
-            reqwest::Method::GET,
-            "/admin/permissions",
-            None,
-        )
-        .await
-        {
-            Ok(response) => response,
-            Err(error) => error.into_response(),
-        };
-    }
-
-    match state
-        .runtime
-        .with_schema_manager(|schema_manager| schema_manager.current_permissions())
-    {
-        Ok(current) => (
-            StatusCode::OK,
-            Json(match current {
-                Some(current) => StoredPermissionsResponse {
-                    head: Some(permissions_head_view(current.head)),
-                    permissions: Some(permissions_map_view(current.permissions)),
-                },
-                None => StoredPermissionsResponse {
-                    head: None,
-                    permissions: None,
-                },
-            }),
-        )
-            .into_response(),
-        Err(err) => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ErrorResponse::internal(format!(
-                "failed to read current permissions: {err}"
-            ))),
-        )
-            .into_response(),
-    }
-}
-
-async fn publish_permissions_handler(
-    State(state): State<Arc<ServerState>>,
-    headers: HeaderMap,
-    Json(request): Json<PublishPermissionsRequest>,
-) -> impl IntoResponse {
-    let admin_secret = headers
-        .get("X-Jazz-Admin-Secret")
-        .and_then(|v| v.to_str().ok());
-
-    match validate_admin_secret(admin_secret, &state.auth_config) {
-        Ok(()) => {}
-        Err((status, msg)) => {
-            return (status, Json(ErrorResponse::unauthorized(msg))).into_response();
-        }
-    }
-
-    if matches!(
-        &state.catalogue_authority,
-        CatalogueAuthorityMode::Forward { .. }
-    ) {
-        let body = match serde_json::to_vec(&request) {
-            Ok(body) => body,
-            Err(err) => {
-                return (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    Json(ErrorResponse::internal(format!(
-                        "failed to serialize permissions publish request: {err}"
-                    ))),
-                )
-                    .into_response();
-            }
-        };
-        return match forward_catalogue_request(
-            &state,
-            reqwest::Method::POST,
-            "/admin/permissions",
-            Some(body),
-        )
-        .await
-        {
-            Ok(response) => response,
-            Err(error) => error.into_response(),
-        };
-    }
-
-    let schema_hash = match parse_schema_hash_param(&request.schema_hash) {
-        Ok(hash) => hash,
-        Err(message) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ErrorResponse::bad_request(message)),
-            )
-                .into_response();
-        }
-    };
-
-    let expected_parent_bundle_object_id = match request.expected_parent_bundle_object_id {
-        Some(object_id) => match parse_object_id_param(&object_id) {
-            Ok(object_id) => Some(object_id),
-            Err(message) => {
-                return (
-                    StatusCode::BAD_REQUEST,
-                    Json(ErrorResponse::bad_request(message)),
-                )
-                    .into_response();
-            }
-        },
-        None => None,
-    };
-
-    match state.runtime.known_schema(&schema_hash) {
-        Ok(Some(_)) => {}
-        Ok(None) => {
-            return (
-                StatusCode::NOT_FOUND,
-                Json(ErrorResponse::not_found(format!(
-                    "target schema catalogue not found for hash {}",
-                    schema_hash
-                ))),
-            )
-                .into_response();
-        }
-        Err(err) => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::internal(format!(
-                    "failed to read known schemas: {err}"
-                ))),
-            )
-                .into_response();
-        }
-    }
-
-    let permissions = request
-        .permissions
-        .into_iter()
-        .map(|(table_name, policies)| (TableName::new(table_name), policies))
-        .collect();
-
-    match state.runtime.publish_permissions_bundle(
-        schema_hash,
-        permissions,
-        expected_parent_bundle_object_id,
-    ) {
-        Ok(_) => match state.runtime.with_schema_manager(|schema_manager| {
-            schema_manager
-                .current_permissions_head()
-                .map(permissions_head_view)
-        }) {
-            Ok(head) => {
-                (StatusCode::CREATED, Json(PermissionsHeadResponse { head })).into_response()
-            }
-            Err(err) => (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::internal(format!(
-                    "failed to read published permissions head: {err}"
-                ))),
-            )
-                .into_response(),
-        },
-        Err(crate::runtime_tokio::RuntimeError::WriteError(message))
-            if message.starts_with("stale permissions parent") =>
-        {
-            (
-                StatusCode::CONFLICT,
-                Json(ErrorResponse::bad_request(message)),
-            )
-                .into_response()
-        }
-        Err(err) => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ErrorResponse::internal(format!(
-                "failed to publish permissions catalogue: {err}"
-            ))),
-        )
-            .into_response(),
-    }
-}
-
-/// Publish a reviewed migration edge into the catalogue.
-///
-/// Requires a valid admin secret. The source and target schemas must already be
-/// known to the server; only the lens edge itself is created here.
-async fn publish_migration_handler(
-    State(state): State<Arc<ServerState>>,
-    headers: HeaderMap,
-    Json(request): Json<PublishMigrationRequest>,
-) -> impl IntoResponse {
-    let admin_secret = headers
-        .get("X-Jazz-Admin-Secret")
-        .and_then(|v| v.to_str().ok());
-
-    match validate_admin_secret(admin_secret, &state.auth_config) {
-        Ok(()) => {}
-        Err((status, msg)) => {
-            return (status, Json(ErrorResponse::unauthorized(msg))).into_response();
-        }
-    }
-
-    if matches!(
-        &state.catalogue_authority,
-        CatalogueAuthorityMode::Forward { .. }
-    ) {
-        let body = match serde_json::to_vec(&request) {
-            Ok(body) => body,
-            Err(err) => {
-                return (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    Json(ErrorResponse::internal(format!(
-                        "failed to serialize migration publish request: {err}"
-                    ))),
-                )
-                    .into_response();
-            }
-        };
-        return match forward_catalogue_request(
-            &state,
-            reqwest::Method::POST,
-            "/admin/migrations",
-            Some(body),
-        )
-        .await
-        {
-            Ok(response) => response,
-            Err(error) => error.into_response(),
-        };
-    }
-
-    let source_hash = match parse_schema_hash_param(&request.from_hash) {
-        Ok(hash) => hash,
-        Err(message) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ErrorResponse::bad_request(message)),
-            )
-                .into_response();
-        }
-    };
-
-    let target_hash = match parse_schema_hash_param(&request.to_hash) {
-        Ok(hash) => hash,
-        Err(message) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ErrorResponse::bad_request(message)),
-            )
-                .into_response();
-        }
-    };
-
-    let source_schema = match state.runtime.known_schema(&source_hash) {
-        Ok(Some(schema)) => schema,
-        Ok(None) => {
-            return (
-                StatusCode::NOT_FOUND,
-                Json(ErrorResponse::not_found(format!(
-                    "source schema catalogue not found for hash {}",
-                    source_hash
-                ))),
-            )
-                .into_response();
-        }
-        Err(err) => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::internal(format!(
-                    "failed to read source schema catalogue: {err}"
-                ))),
-            )
-                .into_response();
-        }
-    };
-
-    let target_schema = match state.runtime.known_schema(&target_hash) {
-        Ok(Some(schema)) => schema,
-        Ok(None) => {
-            return (
-                StatusCode::NOT_FOUND,
-                Json(ErrorResponse::not_found(format!(
-                    "target schema catalogue not found for hash {}",
-                    target_hash
-                ))),
-            )
-                .into_response();
-        }
-        Err(err) => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::internal(format!(
-                    "failed to read target schema catalogue: {err}"
-                ))),
-            )
-                .into_response();
-        }
-    };
-
-    let mut forward = LensTransform::new();
-    for table_lens in request.forward {
-        let table_name = table_lens.table;
-        if table_lens.added && table_lens.removed {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ErrorResponse::bad_request(format!(
-                    "table {} cannot be both added and removed",
-                    table_name
-                ))),
-            )
-                .into_response();
-        }
-        if (table_lens.added || table_lens.removed) && table_lens.renamed_from.is_some() {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ErrorResponse::bad_request(format!(
-                    "table {} cannot combine added/removed markers with renamedFrom",
-                    table_name
-                ))),
-            )
-                .into_response();
-        }
-        if (table_lens.added || table_lens.removed) && !table_lens.operations.is_empty() {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ErrorResponse::bad_request(format!(
-                    "table {} cannot combine added/removed markers with column operations",
-                    table_name
-                ))),
-            )
-                .into_response();
-        }
-        if table_lens.added {
-            let target_table_name = TableName::from(table_name.clone());
-            let schema = match target_schema.get(&target_table_name) {
-                Some(schema) => schema.clone(),
-                None => {
-                    return (
-                        StatusCode::BAD_REQUEST,
-                        Json(ErrorResponse::bad_request(format!(
-                            "createTables references unknown target table {}",
-                            table_name
-                        ))),
-                    )
-                        .into_response();
-                }
-            };
-            forward.push(
-                LensOp::AddTable {
-                    table: table_name.clone(),
-                    schema,
-                },
-                false,
-            );
-        }
-        if table_lens.removed {
-            let source_table_name = TableName::from(table_name.clone());
-            let schema = match source_schema.get(&source_table_name) {
-                Some(schema) => schema.clone(),
-                None => {
-                    return (
-                        StatusCode::BAD_REQUEST,
-                        Json(ErrorResponse::bad_request(format!(
-                            "dropTables references unknown source table {}",
-                            table_name
-                        ))),
-                    )
-                        .into_response();
-                }
-            };
-            forward.push(
-                LensOp::RemoveTable {
-                    table: table_name.clone(),
-                    schema,
-                },
-                false,
-            );
-        }
-        if let Some(renamed_from) = table_lens.renamed_from {
-            forward.push(
-                LensOp::RenameTable {
-                    old_name: renamed_from,
-                    new_name: table_name.clone(),
-                },
-                false,
-            );
-        }
-        for operation in table_lens.operations {
-            let op = match operation {
-                PublishLensOp::Introduce {
-                    column,
-                    column_type,
-                    value,
-                } => LensOp::AddColumn {
-                    table: table_name.clone(),
-                    column,
-                    column_type,
-                    default: value,
-                },
-                PublishLensOp::Drop {
-                    column,
-                    column_type,
-                    value,
-                } => LensOp::RemoveColumn {
-                    table: table_name.clone(),
-                    column,
-                    column_type,
-                    default: value,
-                },
-                PublishLensOp::Rename { column, value } => LensOp::RenameColumn {
-                    table: table_name.clone(),
-                    old_name: column,
-                    new_name: value,
-                },
-            };
-            forward.push(op, false);
-        }
-    }
-
-    let lens = Lens::new(source_hash, target_hash, forward);
-    let object_id = match state.runtime.publish_lens(&lens) {
-        Ok(object_id) => object_id,
-        Err(err) => {
-            return (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::internal(format!(
-                    "failed to publish migration lens: {err}"
-                ))),
-            )
-                .into_response();
-        }
-    };
-
-    if let Err(err) = state.runtime.flush().await {
-        return (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ErrorResponse::internal(format!(
-                "failed to flush published migration lens: {err}"
-            ))),
-        )
-            .into_response();
-    }
-
-    (
-        StatusCode::CREATED,
-        Json(PublishMigrationResponse {
-            object_id: object_id.to_string(),
-            from_hash: request.from_hash,
-            to_hash: request.to_hash,
-        }),
-    )
-        .into_response()
-}
-
-async fn admin_subscription_introspection_handler(
-    State(state): State<Arc<ServerState>>,
-    headers: HeaderMap,
-    Query(params): Query<AdminSubscriptionIntrospectionParams>,
-) -> impl IntoResponse {
-    let admin_secret = headers
-        .get("X-Jazz-Admin-Secret")
-        .and_then(|v| v.to_str().ok());
-
-    match validate_admin_secret(admin_secret, &state.auth_config) {
-        Ok(()) => {}
-        Err((status, msg)) => {
-            return (status, Json(ErrorResponse::unauthorized(msg))).into_response();
-        }
-    }
-
-    let Some(app_id_text) = params.app_id.as_deref() else {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(ErrorResponse::bad_request(
-                "appId query parameter is required",
-            )),
-        )
-            .into_response();
-    };
-
-    let requested_app_id = match parse_app_id_param(app_id_text) {
-        Ok(app_id) => app_id,
-        Err(message) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(ErrorResponse::bad_request(message)),
-            )
-                .into_response();
-        }
-    };
-
-    if requested_app_id != state.app_id {
-        return (
-            StatusCode::NOT_FOUND,
-            Json(ErrorResponse::not_found(format!(
-                "app not found: {}",
-                app_id_text.trim()
-            ))),
-        )
-            .into_response();
-    }
-
-    match state.runtime.server_subscription_telemetry() {
-        Ok(queries) => Json(AdminSubscriptionIntrospectionResponse {
-            app_id: state.app_id.to_string(),
-            generated_at: unix_timestamp_millis(),
-            queries,
-        })
-        .into_response(),
-        Err(err) => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ErrorResponse::internal(format!(
-                "failed to read subscription telemetry: {err}"
-            ))),
-        )
-            .into_response(),
-    }
-}
-
-fn parse_schema_hash_param(hash_text: &str) -> Result<SchemaHash, String> {
-    let decoded_hash_bytes = hex::decode(hash_text)
-        .map_err(|_| "invalid schema hash: expected hex string".to_string())?;
-    if decoded_hash_bytes.len() != 32 {
-        return Err("invalid schema hash: expected 64 hex chars".to_string());
-    }
-
-    let mut hash_bytes = [0u8; 32];
-    hash_bytes.copy_from_slice(&decoded_hash_bytes);
-    Ok(SchemaHash::from_bytes(hash_bytes))
-}
-
-fn connection_schema_diagnostics_from_handshake(
-    state: &Arc<ServerState>,
-    handshake: &crate::transport_manager::AuthHandshake,
-) -> Result<
-    Option<crate::sync_manager::ConnectionSchemaDiagnostics>,
-    crate::runtime_tokio::RuntimeError,
-> {
-    let Some(client_schema_hash) = handshake.declared_schema_hash() else {
-        return Ok(None);
-    };
-
-    let diagnostics = state
-        .runtime
-        .with_schema_manager(|sm| sm.connection_schema_diagnostics(client_schema_hash))?;
-    Ok(diagnostics.has_issues().then_some(diagnostics))
-}
-
-fn parse_object_id_param(object_id_text: &str) -> Result<ObjectId, String> {
-    let uuid = Uuid::parse_str(object_id_text)
-        .map_err(|_| "invalid object id: expected UUID".to_string())?;
-    Ok(ObjectId::from_uuid(uuid))
-}
-
-fn parse_app_id_param(app_id_text: &str) -> Result<AppId, String> {
-    let trimmed = app_id_text.trim();
-    if trimmed.is_empty() {
-        return Err("invalid appId: expected UUID or app identifier".to_string());
-    }
-
-    if let Ok(app_id) = AppId::from_string(trimmed) {
-        return Ok(app_id);
-    }
-
-    if trimmed
-        .chars()
-        .all(|char| char.is_ascii_alphanumeric() || matches!(char, '-' | '_' | '.'))
-    {
-        return Ok(AppId::from_name(trimmed));
-    }
-
-    Err("invalid appId: expected UUID or app identifier".to_string())
-}
-
-fn permissions_head_view(
-    head: crate::schema_manager::manager::PermissionsHeadSummary,
-) -> PermissionsHeadView {
-    PermissionsHeadView {
-        schema_hash: head.schema_hash.to_string(),
-        version: head.version,
-        parent_bundle_object_id: head
-            .parent_bundle_object_id
-            .map(|object_id| object_id.to_string()),
-        bundle_object_id: head.bundle_object_id.to_string(),
-    }
-}
-
-fn permissions_map_view(
-    permissions: std::collections::HashMap<TableName, TablePolicies>,
-) -> std::collections::HashMap<String, TablePolicies> {
-    permissions
-        .into_iter()
-        .map(|(table_name, policies)| (table_name.to_string(), policies))
-        .collect()
-}
-
-fn unix_timestamp_millis() -> u64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_millis()
-        .min(u128::from(u64::MAX)) as u64
-}
-
-/// Health check endpoint.
-async fn health_handler() -> impl IntoResponse {
-    Json(serde_json::json!({
-        "status": "healthy"
-    }))
-}
-
-// ============================================================================
-// WebSocket transport — Task 9
-// ============================================================================
-
-/// WebSocket upgrade endpoint.
-///
-/// Clients send an `AuthHandshake` binary frame (4-byte length prefix + JSON),
-/// receive a `ConnectedResponse` frame, then exchange binary frames
-/// bidirectionally until the connection closes.
-pub async fn ws_handler(
-    ws: WebSocketUpgrade,
-    State(state): State<Arc<ServerState>>,
-    headers: HeaderMap,
-) -> Response {
-    ws.on_upgrade(move |socket| handle_ws_connection(socket, state, headers))
-}
-
-/// Outcome of authenticating a WS handshake — mirrors `ClientSetup` in
-/// the old `/sync` + `/events` handlers.
-#[derive(Debug)]
-enum WsClientSetup {
-    Backend,
-    Session(crate::query_manager::session::Session),
-}
-
-/// Authenticate a WebSocket `AuthHandshake`.
-///
-/// Priority is:
-/// 1. `admin_secret` valid → `WsClientSetup::Backend`
-/// 2. `backend_secret` present + no session header → `WsClientSetup::Backend`
-/// 3. Otherwise → `extract_session` → `WsClientSetup::Session`
-///
-/// Returns `Err(message)` on auth failure; the caller should send a
-/// `ServerEvent::Error` frame before closing.
-async fn authenticate_ws_handshake(
-    handshake: &crate::transport_manager::AuthHandshake,
-    request_headers: &HeaderMap,
-    state: &Arc<ServerState>,
-) -> Result<WsClientSetup, String> {
-    use axum::http::HeaderValue;
-    use base64::Engine as _;
-
-    let auth = &handshake.auth;
-
-    // `admin_secret` is an explicit request to run this WS transport as the
-    // backend. Validate it first and short-circuit all user-scoped auth.
-    if let Some(admin_secret) = auth.admin_secret.as_deref() {
-        validate_admin_secret(Some(admin_secret), &state.auth_config)
-            .map_err(|(_, msg)| msg.to_string())?;
-        return Ok(WsClientSetup::Backend);
-    }
-
-    if request_uses_cookie_auth(handshake, request_headers, &state.auth_config) {
-        validate_ws_cookie_origin(request_headers)?;
-    }
-
-    // Build a synthetic HeaderMap from the handshake auth fields, layered on
-    // top of the original upgrade request so cookie-based auth remains visible.
-    let mut headers = request_headers.clone();
-
-    if let Some(jwt) = &auth.jwt_token {
-        let value = HeaderValue::from_str(&format!("Bearer {jwt}"))
-            .map_err(|e| format!("invalid jwt_token header value: {e}"))?;
-        headers.insert(axum::http::header::AUTHORIZATION, value);
-    }
-    if let Some(secret) = &auth.backend_secret {
-        let value = HeaderValue::from_str(secret)
-            .map_err(|e| format!("invalid backend_secret header value: {e}"))?;
-        headers.insert("X-Jazz-Backend-Secret", value);
-    }
-    if let Some(session_val) = &auth.backend_session {
-        let json = serde_json::to_string(session_val)
-            .map_err(|e| format!("failed to serialise backend_session: {e}"))?;
-        let b64 = base64::engine::general_purpose::STANDARD.encode(json.as_bytes());
-        let value = HeaderValue::from_str(&b64)
-            .map_err(|e| format!("invalid backend_session header value: {e}"))?;
-        headers.insert("X-Jazz-Session", value);
-    }
-
-    let has_jwt = headers.get(axum::http::header::AUTHORIZATION).is_some();
-    let has_session_header = headers.get("X-Jazz-Session").is_some();
-    let backend_secret = headers
-        .get("X-Jazz-Backend-Secret")
-        .and_then(|v| v.to_str().ok());
-
-    // 2. Backend secret — only when no user-scoped JWT is present.  Clients
-    //    that carry both a backend_secret and a jwt_token (e.g. test helpers
-    //    that mirror the full credential set) must be treated as users so the
-    //    connection carries a session for row-level policy evaluation.
-    if backend_secret.is_some() && !has_jwt && !has_session_header {
-        validate_backend_secret(backend_secret, &state.auth_config)
-            .map_err(|(_, msg)| msg.to_string())?;
-        return Ok(WsClientSetup::Backend);
-    }
-
-    // 3. JWT / session-impersonation path.
-    let session = extract_session(
-        &headers,
-        state.app_id,
-        &state.auth_config,
-        state.jwt_verifier.as_deref(),
-    )
-    .await
-    .map_err(|e| serde_json::to_string(&e).unwrap_or_else(|_| "authentication failed".into()))?;
-
-    let session =
-        session.ok_or_else(|| "Session required. Provide JWT or backend secret.".to_string())?;
-
-    Ok(WsClientSetup::Session(session))
-}
-
-fn request_uses_cookie_auth(
-    handshake: &crate::transport_manager::AuthHandshake,
-    request_headers: &HeaderMap,
-    auth_config: &crate::middleware::AuthConfig,
-) -> bool {
-    let Some(cookie_name) = auth_config.auth_cookie_name.as_deref() else {
-        return false;
-    };
-
-    let has_explicit_auth = handshake.auth.jwt_token.is_some()
-        || handshake.auth.backend_secret.is_some()
-        || handshake.auth.backend_session.is_some()
-        || handshake.auth.admin_secret.is_some()
-        || request_headers
-            .get(axum::http::header::AUTHORIZATION)
-            .is_some()
-        || request_headers.get("X-Jazz-Backend-Secret").is_some()
-        || request_headers.get("X-Jazz-Session").is_some()
-        || request_headers.get("X-Jazz-Admin-Secret").is_some();
-
-    if has_explicit_auth {
-        return false;
-    }
-
-    request_cookie_value(request_headers, cookie_name).is_some()
-}
-
-fn request_cookie_value<'a>(headers: &'a HeaderMap, name: &str) -> Option<&'a str> {
-    let cookie_header = headers
-        .get(axum::http::header::COOKIE)
-        .and_then(|value| value.to_str().ok())?;
-
-    cookie_header.split(';').find_map(|segment| {
-        let trimmed = segment.trim();
-        let (candidate_name, candidate_value) = trimmed.split_once('=')?;
-        if candidate_name == name && !candidate_value.is_empty() {
-            Some(candidate_value)
-        } else {
-            None
-        }
-    })
-}
-
-fn validate_ws_cookie_origin(headers: &HeaderMap) -> Result<(), String> {
-    let host = headers
-        .get("X-Forwarded-Host")
-        .or_else(|| headers.get(axum::http::header::HOST))
-        .and_then(|value| value.to_str().ok())
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .ok_or_else(|| "Cookie auth requires Host header".to_string())?;
-
-    let origin = headers
-        .get(axum::http::header::ORIGIN)
-        .and_then(|value| value.to_str().ok())
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .ok_or_else(|| "Cookie auth requires Origin header".to_string())?;
-
-    let origin_uri: axum::http::Uri = origin
-        .parse()
-        .map_err(|_| "Cookie auth requires a valid Origin header".to_string())?;
-    let origin_authority = origin_uri
-        .authority()
-        .map(|authority| authority.as_str())
-        .ok_or_else(|| "Cookie auth requires an Origin authority".to_string())?;
-
-    let is_allowed_origin = origin_authority.eq_ignore_ascii_case(host)
-        || is_loopback_cookie_origin(origin_uri.scheme_str(), origin_authority, host)?;
-
-    if is_allowed_origin {
-        Ok(())
-    } else {
-        Err("Cookie auth Origin must match Host".to_string())
-    }
-}
-
-fn is_loopback_cookie_origin(
-    origin_scheme: Option<&str>,
-    origin_authority: &str,
-    host: &str,
-) -> Result<bool, String> {
-    if !matches!(origin_scheme, Some("http") | Some("https")) {
-        return Ok(false);
-    }
-
-    let origin_authority: axum::http::uri::Authority = origin_authority
-        .parse()
-        .map_err(|_| "Cookie auth requires a valid Origin authority".to_string())?;
-    let host_authority: axum::http::uri::Authority = host
-        .parse()
-        .map_err(|_| "Cookie auth requires a valid Host header".to_string())?;
-
-    Ok(
-        is_loopback_dev_host(origin_authority.host())
-            && is_loopback_dev_host(host_authority.host()),
-    )
-}
-
-fn is_loopback_dev_host(host: &str) -> bool {
-    let host = host.trim_matches(['[', ']']);
-    host.eq_ignore_ascii_case("localhost")
-        || host.to_ascii_lowercase().ends_with(".localhost")
-        || host == "127.0.0.1"
-        || host == "::1"
-}
-
-/// Send a `ServerEvent::Error` frame on the socket, best-effort.
-async fn send_ws_error(socket: &mut WebSocket, message: &str) {
-    use crate::jazz_transport::ErrorCode;
-    let event = crate::jazz_transport::ServerEvent::Error {
-        message: message.to_string(),
-        code: ErrorCode::Unauthorized,
-    };
-    if let Ok(bytes) = serde_json::to_vec(&event) {
-        let _ = socket
-            .send(Message::Binary(crate::transport_manager::frame_encode(
-                &bytes,
-            )))
-            .await;
-    }
-}
-
-async fn handle_ws_connection(
-    mut socket: WebSocket,
-    state: Arc<ServerState>,
-    request_headers: HeaderMap,
-) {
-    // 1. Read the first binary frame — expected to be AuthHandshake.
-    let first = match socket.recv().await {
-        Some(Ok(Message::Binary(b))) => b,
-        _ => {
-            let _ = socket.close().await;
-            return;
-        }
-    };
-    let payload = match crate::transport_manager::frame_decode(&first) {
-        Some(p) => p.to_vec(),
-        None => {
-            let _ = socket.close().await;
-            return;
-        }
-    };
-    let handshake =
-        match serde_json::from_slice::<crate::transport_manager::AuthHandshake>(&payload) {
-            Ok(h) => h,
-            Err(_) => {
-                let _ = socket.close().await;
-                return;
-            }
-        };
-
-    // 2. Parse client_id.
-    let client_id = match crate::sync_manager::ClientId::parse(&handshake.client_id) {
-        Some(id) => id,
-        None => {
-            send_ws_error(&mut socket, "missing or invalid client_id").await;
-            let _ = socket.close().await;
-            return;
-        }
-    };
-
-    // 3. Authenticate.
-    let setup = match authenticate_ws_handshake(&handshake, &request_headers, &state).await {
-        Ok(s) => s,
-        Err(msg) => {
-            send_ws_error(&mut socket, &msg).await;
-            let _ = socket.close().await;
-            return;
-        }
-    };
-    let role = match &setup {
-        WsClientSetup::Backend => "backend",
-        WsClientSetup::Session(_) => "session",
-    };
-
-    // 4. Register with ConnectionEventHub (mirrors events_handler).
-    let connection_id = state
-        .next_connection_id
-        .fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-    let (next_sync_seq, mut sync_rx) = state
-        .connection_event_hub
-        .register_connection(connection_id, client_id);
-    {
-        let mut connections = state.connections.write().await;
-        connections.insert(connection_id, ConnectionState { client_id });
-    }
-    state.on_client_connected(client_id).await;
-
-    // 5. Ensure the client state in the runtime.
-    match setup {
-        WsClientSetup::Backend => {
-            let _ = state.runtime.ensure_client_as_backend(client_id);
-        }
-        WsClientSetup::Session(session) => {
-            let _ = state.runtime.ensure_client_with_session(client_id, session);
-        }
-    }
-
-    // 5b. Dispatch connection schema diagnostics if client sent a declared schema hash.
-    match connection_schema_diagnostics_from_handshake(&state, &handshake) {
-        Ok(Some(diagnostics)) => {
-            state.connection_event_hub.dispatch_payload(
-                client_id,
-                crate::sync_manager::SyncPayload::ConnectionSchemaDiagnostics(diagnostics),
-            );
-        }
-        Ok(None) => {}
-        Err(err) => {
-            tracing::error!(
-                %client_id,
-                declared_schema_hash = ?handshake.declared_schema_hash,
-                "failed to compute connection schema diagnostics: {err}"
-            );
-        }
-    }
-
-    // 6. Send the Connected response.
-    let resp = crate::transport_manager::ConnectedResponse {
-        connection_id: connection_id.to_string(),
-        client_id: client_id.to_string(),
-        next_sync_seq: Some(next_sync_seq),
-        catalogue_state_hash: state.runtime.catalogue_state_hash().ok(),
-    };
-    let resp_bytes = match serde_json::to_vec(&resp) {
-        Ok(b) => b,
-        Err(_) => {
-            ws_cleanup(&state, connection_id, client_id).await;
-            let _ = socket.close().await;
-            return;
-        }
-    };
-    if socket
-        .send(Message::Binary(crate::transport_manager::frame_encode(
-            &resp_bytes,
-        )))
-        .await
-        .is_err()
-    {
-        ws_cleanup(&state, connection_id, client_id).await;
-        return;
-    }
-    tracing::info!(connection_id, %client_id, role, "ws client connected");
-
-    // 7. Bidirectional loop: inbound frames from client + outbound updates from hub.
-    //    Also fires a periodic heartbeat so idle connections don't look half-open.
-    let mut heartbeat = tokio::time::interval(std::time::Duration::from_secs(30));
-    // Don't emit a heartbeat immediately after Connected — wait a full tick.
-    heartbeat.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Delay);
-    heartbeat.tick().await; // consume the immediate first tick
-    loop {
-        tokio::select! {
-            msg = socket.recv() => match msg {
-                Some(Ok(Message::Binary(data))) => {
-                    let Some(inner) = crate::transport_manager::frame_decode(&data) else {
-                        continue;
-                    };
-                    let inner = inner.to_vec();
-                    if let Err(e) = state.process_ws_client_frame(client_id, &inner).await {
-                        tracing::warn!(error = ?e, "ws client frame rejected");
-                    }
-                }
-                Some(Ok(Message::Close(_))) | None => break,
-                _ => continue,
-            },
-            update = sync_rx.recv() => {
-                let Some(u) = update else { break };
-                let event = crate::jazz_transport::ServerEvent::SyncUpdate {
-                    seq: Some(u.seq),
-                    payload: Box::new(u.payload),
-                };
-                let bytes = match serde_json::to_vec(&event) {
-                    Ok(b) => b,
-                    Err(_) => continue,
-                };
-                if socket
-                    .send(Message::Binary(
-                        crate::transport_manager::frame_encode(&bytes),
-                    ))
-                    .await
-                    .is_err()
-                {
-                    break;
-                }
-            }
-            _ = heartbeat.tick() => {
-                let event = crate::jazz_transport::ServerEvent::Heartbeat;
-                let Ok(bytes) = serde_json::to_vec(&event) else { continue };
-                if socket
-                    .send(Message::Binary(
-                        crate::transport_manager::frame_encode(&bytes),
-                    ))
-                    .await
-                    .is_err()
-                {
-                    break;
-                }
-            }
-        }
-    }
-
-    ws_cleanup(&state, connection_id, client_id).await;
-    let _ = socket.close().await;
-}
-
-/// Disconnect cleanup: mirrors the drop path in `events_handler`.
-async fn ws_cleanup(state: &Arc<ServerState>, connection_id: u64, client_id: ClientId) {
-    {
-        let mut connections = state.connections.write().await;
-        connections.remove(&connection_id);
-    }
-    state
-        .connection_event_hub
-        .unregister_connection(connection_id);
-    state.on_connection_closed(client_id).await;
-}
-
 #[cfg(test)]
 mod tests {
+    use super::http::*;
+    use super::utils::*;
+    use super::websocket::*;
     use super::*;
+
+    use axum::extract::{Path, Query};
+    use axum::http::{HeaderMap, StatusCode};
+    use axum::response::Json;
+
+    use uuid::Uuid;
+
+    use crate::object::ObjectId;
+    use crate::query_manager::types::{SchemaHash, TableName};
+    use crate::schema_manager::{AppId, LensOp};
+    use crate::server::ConnectionState;
 
     use std::collections::BTreeMap;
     use std::fmt;
